+++ src/lib/theme/colors.ts
@@ -157,6 +157,10 @@ const colors = {
   offWhite: '#F8FAFC',
   black: '#141414',
   offBlack: '#0F172A',
+  spaceGray: '#141414',
+  asteroidBeige: '#f8f8f7',
+  lightBlue: '#DFFBFF',
+  lightPurple: '#DDD6FE',
   transparent: 'transparent',
   danger: getColor('red', 700),
   success: getColor('green', 200),
